sideBar button action connect to api

+++ client/src/Lobby/ChatList.tsx
@@ -4,6 +4,7 @@ import useSwr from 'swr';
 import lock from '../Images/lock.png'
 import unlock from '../Images/unlock.png'
 import "./styles/ChatList.scss"
+import PwdCheckModal from "./PwdCheckModal";
 
 const ChatList = () => {
 	const fetcher = async (url:string) => {
@@ -13,12 +14,26 @@ const ChatList = () => {
 
 	const {data, error} = useSwr<{title:string, num:number, security:string}[]>('Lobby/chatList', fetcher)
 
+	const [ClickedRoomTitle, setClickedRoomTitle] = useState('')
+	const [PwdCheckModalState, setPwdCheckModalState] = useState(false);
+
+	const openPwdCheckModal = (title:string) => {
+		setPwdCheckModalState(true);
+		setClickedRoomTitle(title)
+	}
+	const closePwdCheckModal = () => {
+		setPwdCheckModalState(false);
+	}
+
 	return (
 	<>
 	{data?.map(chatRoom=>
 		<button type="button" className="chatList" onClick={() => {
-			document.location.href = '/Chat';
-			sessionStorage.setItem('roomName', chatRoom.title)
+			if (chatRoom.security === 'protected')
+				openPwdCheckModal(chatRoom.title)
+			else
+				document.location.href = '/chat'
+				sessionStorage.setItem('roomName', chatRoom.title)
 		}}>
 			<span className="chatList-left">
 				<div className="chatList-num">{chatRoom.num}</div>
@@ -29,8 +44,10 @@ const ChatList = () => {
 			<span className="roomName">
 				{chatRoom.title}
 			</span>
+
 		</button>
 	)}
+	<PwdCheckModal open={PwdCheckModalState} close={closePwdCheckModal} chatRoomID={ClickedRoomTitle}></PwdCheckModal>
 	</>
 	)
 }
